Split `BufferReader` and `BufferWriter` string serialization logic into partial classes for .NET-specific implementations.
