+++ LiteDB/LiteDB.csproj
@@ -43,6 +43,17 @@
   <PropertyGroup Condition="'$(TargetFramework)' == 'net8.0'">
     <DefineConstants>HAVE_SHA1_MANAGED;HAVE_APP_DOMAIN;HAVE_PROCESS;HAVE_ENVIRONMENT</DefineConstants>
   </PropertyGroup>
+    
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <Compile Remove="**\*.NetCore.cs" />
+    <None Remove="**\*.NetCore.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
+    <Compile Remove="**\*.NetStd.cs" />
+    <None Remove="**\*.NetStd.cs" />
+  </ItemGroup>
 
   <!-- Begin References -->
   <ItemGroup>
@@ -50,8 +61,9 @@
     <None Include="..\icon_64x64.png" Pack="true" PackagePath="\" />
   </ItemGroup>
 
-  <ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
     <PackageReference Include="System.Buffers" Version="4.5.1" />
+    <PackageReference Include="System.Memory" Version="4.5.0" />
   </ItemGroup>
   
   <!-- End References -->
